Use YOCaml built-in server and remove collection
This patch uses YOCaml's brand new `built-in' server. It is likely that
the latter is neither efficient nor scalable, but it has the merit of
not being written in Python. In addition to not being efficient, the
server is very slow because, with each request (not leading to a 404),
it executes the YOCaml program that allows to build the site... Pffrt.

The big advantage of this approach is that you can write your
article (or recipe ;)) as you go along, without having to worry about
restarting the sad Python server. And since YOCaml is smart (coupled
with your article production frequency ;)), it shouldn't be a hindrance.

A good practice for building the blog would be to use a local switch:

```
opam switch create . ocaml-base-compiler.4.13.1
make install-deps
```

See ya.